+++ composer.json
@@ -24,14 +24,15 @@
         "illuminate/events": "5.1.*"
     },
     "require-dev": {
+        "ext-mysqli": "*",
         "illuminate/database": "5.1.*",
         "phpunit/phpunit": "~4.0",
-        "satooshi/php-coveralls": "dev-master",
         "mockery/mockery": "0.9.*"
     },
     "autoload": {
         "psr-4": {
-            "Culpa": "src/"
+            "Culpa\\": "src/",
+            "Culpa\\Tests\\": "tests/"
         }
     },
     "minimum-stability": "dev",
