+++ tests/Bootstrap/Bootstrap.php
@@ -1,0 +1,8 @@
+<?php
+
+$loader = require __DIR__.'/../../vendor/autoload.php';
+
+use Culpa\Tests\Bootstrap\AppFactory;
+use Culpa\Tests\Bootstrap\CulpaTest;
+
+CulpaTest::$app = AppFactory::create();
