+++ tests/BrokenBlameableTest.php
@@ -1,21 +1,18 @@
 <?php
 
-namespace Culpa;
+namespace Culpa\tests;
 
-class BrokenBlameableTest extends \CulpaTest
+use Culpa\Tests\Bootstrap\CulpaTest;
+use Culpa\Tests\Models\BrokenBlameableModel;
+
+class BrokenBlameableTest extends CulpaTest
 {
     private $model;
 
     public function setUp()
     {
-        if (!version_compare(PHP_VERSION, '5.4.0', '>=')) {
-            return $this->markTestSkipped('This test uses a model that uses traits.');
-        }
-
-        require_once __DIR__ . '/Models/BrokenModel.php';
-
         parent::setUp();
-        $this->model = new BrokenBlameableModel;
+        $this->model = new BrokenBlameableModel();
     }
 
     public function testBlameables()
