--- tests/Culpa/Models/CustomFieldsModel.php
@@ -1,1 +1,0 @@
-<?phpnamespace Culpa;use Illuminate\Database\Eloquent\Model;use Illuminate\Support\Facades\Config;/** * A model with custom names for fields * PHP 5.4+ * * Custom fields need to have their own relationship-accessors written */class CustomBlameableModel extends Model{    use Blameable;    protected $table = 'posts';    protected $softDelete = true;    protected $blameable = array(        'created' => 'author_id',        'updated' => 'editor_id',        'deleted' => 'purger_id',    );        public function author()    {        return $this->belongsTo(Config::get('culpa::users.classname', 'User'));    }        public function editor()    {        return $this->belongsTo(Config::get('culpa::users.classname', 'User'));    }        public function purger()    {        return $this->belongsTo(Config::get('culpa::users.classname', 'User'));    }}CustomBlameableModel::observe(new BlameableObserver);
