+++ tests/Models/BrokenModel.php
@@ -1,14 +1,16 @@
 <?php
 
-namespace Culpa;
+namespace Culpa\tests\Models;
 
+use Culpa\Observer\BlameableObserver;
 use Illuminate\Database\Eloquent\Model;
+use Culpa\Model\Blameable;
 
 /**
  * A model with a silly $blameable value
- * PHP 5.4+
+ * PHP 5.4+.
  */
-class BrokenBlameableModel extends Model
+class BrokenModel extends Model
 {
     use Blameable;
     protected $table = 'posts';
@@ -16,4 +18,5 @@ class BrokenBlameableModel extends Model
 
     protected $blameable = 42;
 }
-BrokenBlameableModel::observe(new BlameableObserver);
+
+BrokenBlameableModel::observe(new BlameableObserver());
