+++ tests/NotBlameableTest.php
@@ -1,21 +1,18 @@
 <?php
 
-namespace Culpa;
+namespace Culpa\tests;
 
-class NotBlameableTest extends \CulpaTest
+use Culpa\Tests\Bootstrap\CulpaTest;
+use Culpa\Tests\Models\NotBlameableModel;
+
+class NotBlameableTest extends CulpaTest
 {
     private $model;
 
     public function setUp()
     {
-        if (!version_compare(PHP_VERSION, '5.4.0', '>=')) {
-            return $this->markTestSkipped('This test uses a model that uses traits.');
-        }
-
-        require_once __DIR__ . '/Models/NotBlameableModel.php';
-
         parent::setUp();
-        $this->model = new NotBlameableModel;
+        $this->model = new NotBlameableModel();
     }
 
     public function testBlameables()
